# Fixed #32948 -- Optimized django.utils.tree.Node and its subclasses.