--- a/django/contrib/contenttypes/fields.py
+++ b/django/contrib/contenttypes/fields.py
@@ -627,17 +627,19 @@ def create_generic_related_manager(superclass, rel):
             queryset._add_hints(instance=instances[0])
             queryset = queryset.using(queryset._db or self._db)
             # Group instances by content types.
-            content_type_queries = (
-                models.Q(
-                    (f"{self.content_type_field_name}__pk", content_type_id),
-                    (f"{self.object_id_field_name}__in", {obj.pk for obj in objs}),
+            content_type_queries = [
+                models.Q.create(
+                    [
+                        (f"{self.content_type_field_name}__pk", content_type_id),
+                        (f"{self.object_id_field_name}__in", {obj.pk for obj in objs}),
+                    ]
                 )
                 for content_type_id, objs in itertools.groupby(
                     sorted(instances, key=lambda obj: self.get_content_type(obj).pk),
                     lambda obj: self.get_content_type(obj).pk,
                 )
-            )
-            query = models.Q(*content_type_queries, _connector=models.Q.OR)
+            ]
+            query = models.Q.create(content_type_queries, connector=models.Q.OR)
             # We (possibly) need to convert object IDs to the type of the
             # instances' PK in order to match up instances:
             object_id_converter = instances[0]._meta.pk.to_python