--- a/django/db/models/deletion.py
+++ b/django/db/models/deletion.py
@@ -399,9 +399,9 @@ class Collector:
         """
         Get a QuerySet of the related model to objs via related fields.
         """
-        predicate = query_utils.Q(
-            *((f"{related_field.name}__in", objs) for related_field in related_fields),
-            _connector=query_utils.Q.OR,
+        predicate = query_utils.Q.create(
+            [(f"{related_field.name}__in", objs) for related_field in related_fields],
+            connector=query_utils.Q.OR,
         )
         return related_model._base_manager.using(self.using).filter(predicate)