--- a/django/db/models/fields/related.py
+++ b/django/db/models/fields/related.py
@@ -407,12 +407,13 @@ class RelatedField(FieldCacheMixin, Field):
         select all instances of self.related_field.model related through
         this field to obj. obj is an instance of self.model.
         """
-        base_filter = (
-            (rh_field.attname, getattr(obj, lh_field.attname))
-            for lh_field, rh_field in self.related_fields
+        base_q = Q.create(
+            [
+                (rh_field.attname, getattr(obj, lh_field.attname))
+                for lh_field, rh_field in self.related_fields
+            ]
         )
         descriptor_filter = self.get_extra_descriptor_filter(obj)
-        base_q = Q(*base_filter)
         if isinstance(descriptor_filter, dict):
             return base_q & Q(**descriptor_filter)
         elif descriptor_filter: