--- a/django/db/models/fields/related_descriptors.py
+++ b/django/db/models/fields/related_descriptors.py
@@ -1002,19 +1002,21 @@ def create_forward_many_to_many_manager(superclass, rel, reverse):
         do_not_call_in_templates = True
 
         def _build_remove_filters(self, removed_vals):
-            filters = Q((self.source_field_name, self.related_val))
+            filters = Q.create([(self.source_field_name, self.related_val)])
             # No need to add a subquery condition if removed_vals is a QuerySet without
             # filters.
             removed_vals_filters = (
                 not isinstance(removed_vals, QuerySet) or removed_vals._has_filters()
             )
             if removed_vals_filters:
-                filters &= Q((f"{self.target_field_name}__in", removed_vals))
+                filters &= Q.create([(f"{self.target_field_name}__in", removed_vals)])
             if self.symmetrical:
-                symmetrical_filters = Q((self.target_field_name, self.related_val))
+                symmetrical_filters = Q.create(
+                    [(self.target_field_name, self.related_val)]
+                )
                 if removed_vals_filters:
-                    symmetrical_filters &= Q(
-                        (f"{self.source_field_name}__in", removed_vals)
+                    symmetrical_filters &= Q.create(
+                        [(f"{self.source_field_name}__in", removed_vals)]
                     )
                 filters |= symmetrical_filters
             return filters