--- a/django/utils/tree.py
+++ b/django/utils/tree.py
@@ -33,7 +33,7 @@ class Node:
         __init__() with a signature that conflicts with the one defined in
         Node.__init__().
         """
-        obj = Node(children, connector, negated)
+        obj = Node(children, connector or cls.default, negated)
         obj.__class__ = cls
         return obj