--- a/tests/queries/test_q.py
+++ b/tests/queries/test_q.py
@@ -216,6 +216,15 @@ class QTests(SimpleTestCase):
         flatten = list(q.flatten())
         self.assertEqual(len(flatten), 7)
 
+    def test_create_helper(self):
+        items = [("a", 1), ("b", 2), ("c", 3)]
+        for connector in [Q.AND, Q.OR, Q.XOR]:
+            with self.subTest(connector=connector):
+                self.assertEqual(
+                    Q.create(items, connector=connector),
+                    Q(*items, _connector=connector),
+                )
+
 
 class QCheckTests(TestCase):
     def test_basic(self):