--- a/tests/utils_tests/test_tree.py
+++ b/tests/utils_tests/test_tree.py
@@ -69,11 +69,11 @@ class NodeTests(unittest.TestCase):
         self.node1.negate()
         self.assertFalse(self.node1.negated)
 
-    def test_new_instance(self):
+    def test_create(self):
         SubNode = type("SubNode", (Node,), {})
 
         a = SubNode([SubNode(["a", "b"], OR), "c"], AND)
-        b = SubNode._new_instance(a.children, a.connector, a.negated)
+        b = SubNode.create(a.children, a.connector, a.negated)
         self.assertEqual(a, b)
         # Children lists are the same object, but equal.
         self.assertIsNot(a.children, b.children)