ticket-32948

An alternative to #14673 with extra simplifications and optimisations.