@ngnpope 2021-10-03
I ran some basic checks on the performance - seems pretty good!

```bash
$ sudo pyperf system tune

# Before:

$ pyperf timeit --setup 'from django.db.models import Q' --setup 'a = Q(a=1, b=2, c=3)' --setup 'b = Q(d=4, e=5, f=6)' 'a & b'
.....................
Mean +- std dev: 2.46 us +- 0.06 us

$ pyperf timeit --setup 'from django.db.models import Q' --setup 'a = Q(a=1, b=2, c=3)' --setup 'b = Q(d=4, e=5, f=6)' 'a | b'
.....................
Mean +- std dev: 2.79 us +- 0.04 us

$ pyperf timeit --setup 'from django.db.models import Q' --setup 'a = Q(a=1, b=2, c=3)' --setup 'b = Q(d=4, e=5, f=6)' '~a'
.....................
Mean +- std dev: 1.65 us +- 0.02 us

# After:

$ pyperf timeit --setup 'from django.db.models import Q' --setup 'a = Q(a=1, b=2, c=3)' --setup 'b = Q(d=4, e=5, f=6)' 'a & b'
.....................
Mean +- std dev: 1.96 us +- 0.04 us

$ pyperf timeit --setup 'from django.db.models import Q' --setup 'a = Q(a=1, b=2, c=3)' --setup 'b = Q(d=4, e=5, f=6)' 'a | b'
.....................
Mean +- std dev: 2.29 us +- 0.04 us

$ pyperf timeit --setup 'from django.db.models import Q' --setup 'a = Q(a=1, b=2, c=3)' --setup 'b = Q(d=4, e=5, f=6)' '~a'
.....................
Mean +- std dev: 1.00 us +- 0.02 us

$ sudo pyperf system reset
```