@kezabelle 2021-10-05
I still need to come back and sit down and look at the _implementation_ again, but I have finally sat down to compare numbers vs the baseline & the original solution I was proposing (see timeit in ticket), and it's favourable:

Baseline:
```
In [5]: %timeit ~(full | full2)
5.04 µs ± 78 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [6]: %timeit ~(full & full2)
4.83 µs ± 108 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [10]: %timeit reduce(or_, values)
12.1 µs ± 141 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

This branch:
```
In [6]: %timeit ~(full | full2)
3.45 µs ± 33.6 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [7]: %timeit ~(full & full2)
3.18 µs ± 59.3 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

In [9]: %timeit reduce(or_, values)
9.64 µs ± 132 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```
So I definitely think we can run with it.