@ngnpope 2021-10-05
> I still need to come back and sit down and look at the _implementation_ again, ...

Yes - it'll be good to check that I've made reasonable decisions. I have tried to explain in the commit message for each commit why I did things, so that should help with the reviewing.

> ...but I have finally sat down to compare numbers vs the baseline & the original solution I was proposing (see timeit in ticket), and it's favourable:

That's good to know. \o/

Thanks @kezabelle.