@kezabelle 2021-10-19
So, I've sat down and attempted to work through it, trying to remember what I'd originally planned and where you've ended up subsequently. I _think_ it may fine, and probably tidier than the place I ended up; I can't easily reason about the `if` branch changes made in `_combine` because I'm less familiar with what's considered a combinable, and conditional etc. But I trust that it's correct.

Given the unfortunate situation with the attempt to optimise WhereNode cloning (See ticket [33159](https://code.djangoproject.com/ticket/33159), thank goodness for beta testers.) I'm a little cautious of this line precisely because it doesn't copy at all (I think):
```
obj.children = self.children # Don't [:] as .__init__() via .create() does.
```
(caveat: I haven't checked whether you or I introduced it, I may be picking holes in my own proposal for all I now remember.)
I think it'd be good to have that well tested, if it's not already, as it's only a reference right? e.g.: because `add` may `extend` or `append` to `children` we want to be super sure that a later operations on a Q doesn't backfill changes to a previously declared Q which looks untouched. Logic notwithstanding, here's the kind of thing.
```
a = Q(a=1) # this isn't directly modified and looks innocuous.
b = Q(b=2)
c = a & Q(c=3). # a doesn't get mutated here, or by any other operations that can lead to combine/add happening?
```