@ngnpope 2022-07-13
I'll rebase and have another look next week. I think it was basically there, just needed to check whether existing tests covered the expected behaviour well enough.