@ngnpope 2022-07-23
Ok. I've rebased this. To satisfy the concerns around issues with mutating existing objects when copying I added/enhanced some tests to properly capture the existing behaviour to ensure that hasn't changed - basically ensuring that my bullet points in my previous [comment](https://github.com/django/django/pull/14677#issuecomment-946618000) hold true.

Should be good to land.